// src/shared/api/generated/__swagger_client/models/WarehouseDto.ts
/* generated using openapi-typescript-codegen -- do not edit */
/* istanbul ignore file */
/* tslint:disable */
/* eslint-disable */
export type WarehouseDto = {
    id?: string;
    organizationId?: string;
    orgUnitId?: string;
    code?: string;
    name?: string;
    address?: string;
    description?: string;
    active?: boolean;
};